MAR-144: Move models into separate files
