+++ barriers/models/__init__.py
@@ -1,0 +1,7 @@
+from .assessments import Assessment  # noqa
+from .barriers import Barrier  # noqa
+from .companies import Company  # noqa
+from .documents import Document  # noqa
+from .history import HistoryItem  # noqa
+from .notes import Note  # noqa
+from .watchlists import Watchlist  # noqa
